requirements.txt
flask
requests
pytesseract
Pillow
snscrape
python-telegram-bot==13.15
